\docType{data}
\name{attendance}
\alias{atandancd}
\title{Sea lion attendance data}
\description{
   A data.frame of attendance data for 66 female sea lions on a rookery on San Miguel Island, California.
   Each row represents an individual sea lion, while columns hold capture history, year, period, and brand    ID.
}
\author{
  Devin Johnson \email{devin.johnson@noaa.gov}
}
\keyword{data}

