% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{cor.matern}
\alias{cor.matern}
\title{Function to compute Matern correlation (from Hoeting et al. 2005 - Ecological Applications)}
\usage{
cor.matern(theta, D)
}
\arguments{
\item{theta}{A vector containing range, smoothness, and nugget}

\item{D}{Matrix of distances between sites; typically a square matrix with zeros on main diagonal}
}
\value{
A matrix of correlations
}
\description{
Function to compute Matern correlation (from Hoeting et al. 2005 - Ecological Applications)
}
\author{
Andrew Merton
}
\keyword{Matern,}
\keyword{correlation}
