% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_mixture_functions.R
\name{Nmixmcmc}
\alias{Nmixmcmc}
\title{Title MCMC algorithm for fitting simple N-mixture model with varying lambda[i].}
\usage{
Nmixmcmc(Y, q.p, r.p, alpha, beta, n.iter, checkpoint, name, thin)
}
\arguments{
\item{Y}{An n*J matrix of n sites and J repeat visits to each site}

\item{q.p}{shape parameter for beta distribution prior on p}

\item{r.p}{shape parameter for beta distribution prior on p}

\item{alpha}{shape parameter for gamma distribution prior on lambda}

\item{beta}{rate parameter for gamma distribution prior on lambda}

\item{n.iter}{Number of MCMC iterations}

\item{checkpoint}{The number of MCMC iterations that are run between each save of a file containing the MCMC output}

\item{name}{Character name of file output is saved after every checkpoint}

\item{thin}{The thinning rate}
}
\value{
MCMC output
}
\description{
Title MCMC algorithm for fitting simple N-mixture model with varying lambda[i].
}
