% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{plot.prediction.map}
\alias{plot.prediction.map}
\title{plot a map of estimated abundance or related quantity on a grid.}
\usage{
plot.prediction.map(N, Coords, highlight = NULL, myPalette = NULL,
  leg.title = "Abundance")
}
\arguments{
\item{N}{quantity to plot}

\item{Coords}{data.frame holding x and y coordinates of each prediction cell}

\item{highlight}{[optional] A vector of cells to highlight.  Default = NULL}

\item{myPalette}{A prespecified palette from RColorBrewer (default is YlOrBr(9) )}

\item{leg.title}{Title of the legend}
}
\value{
ggplot2 map of predictions on a grid
}
\description{
plot a map of estimated abundance or related quantity on a grid.
}
\author{
Paul Conn \email{paul.conn@noaa.gov}
}
