% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{pois.no.spat}
\alias{pois.no.spat}
\title{JAGS function to conduct a Poisson GLM analysis with no spatial autocorrelation
or overdispersion}
\usage{
pois.no.spat()
}
\value{
RJags object that includes MCMC samples of Poisson regression parameters
}
\description{
JAGS function to conduct a Poisson GLM analysis with no spatial autocorrelation
or overdispersion
}
\author{
Paul B. Conn
}
\keyword{Jags,}
\keyword{MCMC}
\keyword{Poisson,}
