% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{pois.overd.no.spat.cv}
\alias{pois.overd.no.spat.cv}
\title{JAGS function to conduct a Poisson GLM analysis with overdispersion
but no spatial autocorrelation - this is a cleaner version for cross-validation}
\usage{
pois.overd.no.spat.cv()
}
\value{
RJags object that includes MCMC samples of Poisson regression parameters and Gaussian precision for random effects on log scale
}
\description{
JAGS function to conduct a Poisson GLM analysis with overdispersion
but no spatial autocorrelation - this is a cleaner version for cross-validation
}
\author{
Paul B. Conn
}
\keyword{Jags,}
\keyword{MCMC}
\keyword{Poisson,}
