% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{pois.overd.no.spat}
\alias{pois.overd.no.spat}
\title{JAGS function to conduct a Poisson GLM analysis with overdispersion
but no spatial autocorrelation}
\usage{
pois.overd.no.spat()
}
\value{
RJags object that includes MCMC samples of Poisson regression parameters and Gaussian precision for random effects on log scale
}
\description{
JAGS function to conduct a Poisson GLM analysis with overdispersion
but no spatial autocorrelation
}
\author{
Paul B. Conn
}
\keyword{Jags,}
\keyword{MCMC}
\keyword{Poisson,}
