% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{pois.spat.model}
\alias{pois.spat.model}
\title{JAGS function to conduct a Poisson GLM analysis with overdispersion
and spatial autocorrelation}
\usage{
pois.spat.model()
}
\value{
RJags object that includes MCMC samples of Poisson regression parameters,
Gaussian precision for random effects on log scale, and Matern spatial autocorrelation parameters
}
\description{
JAGS function to conduct a Poisson GLM analysis with overdispersion
and spatial autocorrelation
}
\author{
Paul B. Conn
}
\keyword{Jags,}
\keyword{MCMC}
\keyword{Poisson,}
