% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_mixture_functions.R
\name{run.chain.2pl.list}
\alias{run.chain.2pl.list}
\title{Title Wrapper for parallel processing}
\usage{
run.chain.2pl.list(models, Y.list, X, W, n.iter, checkpoint, thin, name.l,
  starting.values)
}
\arguments{
\item{models}{Vector of models to be run}

\item{Y.list}{List of n*J matrices of n sites and J repeat visits to each site}

\item{X}{Covariate matrix for abundance}

\item{W}{Covariate matrix for detection probability}

\item{n.iter}{Number of MCMC iterations}

\item{checkpoint}{The number of MCMC iterations that are run between each save of a file containing the MCMC output}

\item{thin}{The thinning rate}

\item{name.l}{List of character name of file output that is saved after every checkpoint}

\item{starting.values}{A list of MCMC starting values}
}
\value{
MCMC output
}
\description{
Title Wrapper for parallel processing
}
