% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_mixture_functions.R
\name{run.Nmixmcmc.parallel}
\alias{run.Nmixmcmc.parallel}
\title{Title MCMC algorithm for fitting an N-mixture model with covariates in parallel.}
\usage{
run.Nmixmcmc.parallel(Y, X, W, n.iter, checkpoint, thin, name, starting.values)
}
\arguments{
\item{Y}{An n*J matrix of n sites and J repeat visits to each site}

\item{X}{Covariate matrix for abundance}

\item{W}{Covariate matrix for detection probability}

\item{n.iter}{Number of MCMC iterations}

\item{checkpoint}{The number of MCMC iterations that are run between each save of a file containing the MCMC output}

\item{thin}{The thinning rate}

\item{name}{Character name of file output is saved after every checkpoint}

\item{starting.values}{MCMC starting values}
}
\value{
MCMC output
}
\description{
Title MCMC algorithm for fitting an N-mixture model with covariates in parallel.
}
