% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attendance_code.R
\name{run_attendance_analysis}
\alias{run_attendance_analysis}
\title{Open code and run the attendance analysis}
\usage{
run_attendance_analysis(show = TRUE, run = FALSE)
}
\author{
Devin S. Johnson
}
