\docType{data}
\name{SeaOtterData}
\alias{SeaOtterData}
\title{Sea Otter Counts}
\description{
   Counts of sea otters obtainted from temporally replicated
   aerial images taken at 21 locations in Glacier Bay, Alaska..  This dataset consists of 
   \itemize{
   \item{rows} Different spatial locations of aerial photographs in Glacier Bay Alaska, 2016. Values indicate the number of sea otters counted
   \item{columns} Different temporal replicates of aerial photographs of groups of sea otters. 
   }
   See Williams, P.J, M.B. Hooten, J.N. Womble, and M.R. Bower. 2017.
   Estimating occupancy and abundance using aerial images with imperfect
   detection. Methods in Ecology and Evolution  8(12):1679-1689.
}
\author{
  Perry Williams \email{pwill@rams.colostate.edu}
}
\keyword{data}

